+++ components/Board/[id].tsx
@@ -1,0 +1,8 @@
+import { useRouter } from "next/router";
+
+export default function Articles() {
+  const router = useRouter();
+  const { articles } = router.query;
+
+  return <div>article {articles}</div>;
+}
