+++ components/Boards/ArticleSection.module.css
@@ -1,6 +1,7 @@
 .articleSectionHeader {
   display: flex;
   justify-content: space-between;
+  margin-bottom: 24px;
 }
 
 .article {
@@ -27,26 +28,4 @@ height: Fixed (42px)px; */
   line-height: 19.09px;
   text-align: left;
   color: #ffffff;
-}
-
-.searchForm {
-  display: flex;
-  padding: 9px 20px 9px 16px;
-  gap: 10px;
-  border-radius: 12px;
-
-  background: #f3f4f6;
-}
-
-.searchInput {
-  font-family: Pretendard;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 26px;
-  text-align: left;
-  outline: none;
-  width: 400px;
-  background-color: none;
-
-  color: #9ca3af;
 }
