+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@svgr/webpack": "^8.1.0",
+    "axios": "^1.7.4",
     "next": "^13.5.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
